+++ IBAnimatable.playground/Pages/Predefined Animations.xcplaygroundpage/Contents.swift
@@ -1,32 +1,17 @@
 //: ## Predefined Animations
 
 import UIKit
-import XCPlayground
+import PlaygroundSupport
 import IBAnimatable
 
-//: Constants
-let iPhoneWidth = 375
-let iPhoneHeight = 667
-let animatableViewWidth = 100
-let animatableViewX = (iPhoneWidth - animatableViewWidth) / 2
-let animatableViewY = (iPhoneHeight - animatableViewWidth) / 2
-
-//: Set up the iPhone View
-let iPhoneView = UIView(frame: CGRect(x: 0, y: 0, width: iPhoneWidth, height: iPhoneHeight))
-iPhoneView.backgroundColor = .whiteColor()
-XCPlaygroundPage.currentPage.liveView = iPhoneView
+let iPhoneView = PhoneView()
+PlaygroundPage.current.liveView = iPhoneView
 
 //: Set up the animatable View
-let view = AnimatableView(frame: CGRect(x: animatableViewX, y: animatableViewY, width: animatableViewWidth, height: animatableViewWidth))
-view.configAnimatableProperties()
+let view = CircleView()
 iPhoneView.addSubview(view)
 
-view.fillColor = UIColor(red: 0xba/0xff, green: 0x77/0xff, blue: 1, alpha: 1)
-view.borderWidth = 2
-view.borderColor = .purpleColor()
-view.maskType = String(MaskType.Circle)
-
-// For moveTo or moveBy animation
+// For `moveTo` or `moveBy` animation
 view.x = -100
 view.y = 200
 
@@ -34,9 +19,6 @@ view.y = 200
 
 // Uncomment one line to play the animation
 
-//view.moveX()
-//view.moveY()
-//view.moveXY()
 //view.slideInLeft()
 //view.slideInRight()
 //view.slideInDown()
@@ -87,6 +69,8 @@ view.y = 200
 //view.rotate()
 //view.rotate(clockwise: false)
 //view.moveTo()
-view.moveBy()
+//view.moveBy()
+view.zoomInvertIn()
+//view.zoomInvertOut()
 
 //: [Next](@next)