+++ IBAnimatable/FoldAnimator.swift
@@ -36,7 +36,7 @@ public class FoldAnimator: NSObject, AnimatedTransitioning {
     horizontal = fromDirection.isHorizontal
     
     if let firstParam = params.first,
-           unwrappedFolds = Int(firstParam) {
+           let unwrappedFolds = Int(firstParam) {
         self.folds = unwrappedFolds
     }
     
@@ -73,7 +73,7 @@ extension FoldAnimator: UIViewControllerAnimatedTransitioning {
   
   public func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
     let (tempfromView, tempToView, tempContainerView) = getViews(using: transitionContext)
-    guard let fromView = tempfromView, toView = tempToView, containerView = tempContainerView else {
+    guard let fromView = tempfromView, let toView = tempToView, let containerView = tempContainerView else {
       transitionContext.completeTransition(true)
       return
     }